added mapping for retrieving by id

+++ src/main/java/com/interview/todo/repository/TodoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.interview.todo.entity.TodoEntity;
 
-public interface TodoRepository extends CrudRepository<TodoEntity, Short> {
+public interface TodoRepository extends CrudRepository<TodoEntity, Long> {
     
 }
